fix (parser) Changes passages to an object collection
This changes the passages collection inside of `story` to an object
instead of an array. This was causing issues previously where array
index 0 would be undefined and break story data.

BREAKING CHANGE the change from an array to an object requires
migration. If you were iterating over the array, you should now iterate
over the Object.keys() of the object instead.