♻️  Refactor: Adds better tty protection for commitizen
Previously, we relied exclusively on exec < /dev/tty to tell is if we are in a terminal. This works
in most cases, except in instances where we are piping output (release-it is an example). The
POSIX-compatible -t flag allows us to only run the commitizen hook if we're unpiped. Obviously if
you're piping into git commit, you already have a message ready.